requirements.txt
wheel==0.34.2
pika==1.1.0
names==0.3.0
networkx==2.4
pyyaml==5.3.1
tosca-parser==2.0.0
matplotlib==3.2.1